Refactor code for easier readability
